citadel-rs: extract epub behaviour
