+++ src-tauri/src/libs/file_formats/mod.rs
@@ -1,12 +1,11 @@
+use chrono::NaiveDate;
 use std::path::{Path, PathBuf};
 use std::str::FromStr;
 
-use chrono::NaiveDate;
-use epub::doc::EpubDoc;
-
+use super::calibre::ImportableFile;
 use crate::book::{ImportableBookMetadata, ImportableBookType};
 
-use super::calibre::ImportableFile;
+mod epub;
 
 pub enum SupportedFormats {
     EPUB,
@@ -60,9 +59,8 @@ pub fn get_importable_file_metadata(file: ImportableFile) -> Option<ImportableBo
     let format = SupportedFormats::from_file_extension(ext);
 
     match format {
-        Some(SupportedFormats::EPUB) => {
-            let metadata = read_epub_metadata(&file.path);
-            Some(ImportableBookMetadata {
+        Some(SupportedFormats::EPUB) => match epub::read_metadata(&file.path) {
+            Some(metadata) => Some(ImportableBookMetadata {
                 file_type: ImportableBookType::EPUB,
                 title: metadata.title.unwrap_or("".to_string()),
                 author_names: metadata.creator_list,
@@ -76,53 +74,9 @@ pub fn get_importable_file_metadata(file: ImportableFile) -> Option<ImportableBo
                     metadata.publication_date.unwrap_or("".to_string()).as_str(),
                 )
                 .ok(),
-            })
-        }
+            }),
+            _ => None,
+        },
         _ => None,
-    }
-}
-
-pub struct EpubMetadata {
-    pub title: Option<String>,
-    pub creator_list: Option<Vec<String>>,
-    pub identifier: Option<String>,
-    pub publisher: Option<String>,
-    pub publication_date: Option<String>,
-    pub language: Option<String>,
-    pub cover_image_data: Option<Vec<u8>>,
-    pub subjects: Vec<String>,
-}
-
-pub fn cover_data(path: &Path) -> Option<Vec<u8>> {
-    let doc = EpubDoc::new(path);
-    assert!(doc.is_ok());
-    let mut doc = doc.unwrap();
-
-    doc.get_cover().map(|(data, _id)| data)
-}
-
-pub fn read_epub_metadata(path: &Path) -> EpubMetadata {
-    let doc = EpubDoc::new(path);
-    assert!(doc.is_ok());
-    let mut doc = doc.unwrap();
-    let creators = doc
-        .metadata
-        .get("creator")
-        .map(|v| v.to_vec())
-        .unwrap_or(Vec::new());
-
-    EpubMetadata {
-        title: doc.mdata("title"),
-        creator_list: Some(creators),
-        identifier: doc.mdata("identifier"),
-        publisher: doc.mdata("publisher"),
-        language: doc.mdata("language"),
-        cover_image_data: doc.get_cover().map(|(data, _id)| data),
-        publication_date: doc.mdata("date"),
-        subjects: doc
-            .metadata
-            .get("subject")
-            .map(|v| v.to_vec())
-            .unwrap_or(Vec::new()),
     }
 }
